+++ src/Render/interactions/index.ts
@@ -2,5 +2,5 @@ export { enableDrag } from "./drag";
 export { enableZoom } from "./zoom";
 export { enableReferenceLine } from "./referenceLine";
 export { enableDragOutside } from "./dragOutside";
-export { enableSelection } from "./selection";
+export { enableSelection, enableAttractResize } from "./selection";
 export { enableKeyMove } from "./keyMove";